+++ .gitignore
@@ -1,2 +1,3 @@
 data/
+dist/
 __pycache__/
