+++ gllama/util/digest.py
