+++ pyproject.toml
@@ -1,0 +1,22 @@
+[project]
+name = "gllama"
+version = "0.0.1"
+description = "OLlama API \"shell\" wrapper"
+authors = [
+    {name = "Def-Try"},
+    {name = "MichaAI"}
+]
+readme = "README.md"
+requires-python = ">=3.12"
+dependencies = [
+    "aiohttp (>=3.11.12,<4.0.0)",
+    "humanize (>=4.11.0,<5.0.0)"
+]
+
+
+[build-system]
+requires = ["poetry-core>=2.0.0,<3.0.0"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.poetry.scripts]
+gllama = "gllama.__main__:__main__"
